+++ examples/library.rs
@@ -191,9 +191,9 @@ fn main() -> Result<()> {
             .unwrap_or("20")
             .parse::<usize>()?;
         let library: Library<Config, Decoder> = Library::from_config_path(config_path)?;
-        let songs = library.playlist_from::<()>(&[song_path], playlist_length)?;
-        let song_paths = songs
-            .into_iter()
+        let song_paths = library
+            .playlist_from::<()>(&[song_path])?
+            .take(playlist_length)
             .map(|s| s.bliss_song.path.to_string_lossy().to_string())
             .collect::<Vec<String>>();
         for song in song_paths {