+++ examples/library_extra_info.rs
@@ -209,9 +209,9 @@ fn main() -> Result<()> {
             .unwrap_or("20")
             .parse::<usize>()?;
         let library: Library<Config, Decoder> = Library::from_config_path(config_path)?;
-        let songs = library.playlist_from::<ExtraInfo>(&[song_path], playlist_length)?;
-        let playlist = songs
-            .into_iter()
+        let playlist = library
+            .playlist_from::<ExtraInfo>(&[song_path])?
+            .take(playlist_length)
             .map(|s| {
                 (
                     s.bliss_song.path.to_string_lossy().to_string(),